+++ src/line_editor.cpp
@@ -772,6 +772,9 @@ ssize_t LineEditorObject::editInRawMode(ARState &state, RenderingContext &ctx) {
         }
       }
       continue;
+    case EditActionType::COMPLETE_BACKWARD:
+    case EditActionType::PAGER_REVERT:
+      continue; // do nothing (just used in completion pager)
     case EditActionType::BACKWARD_DELETE_CHAR:
       if (ctx.buf.deletePrevChar(nullptr, true)) {
         this->refreshLine(state, ctx);